refactor
https://github.com/metasfresh/metasfresh/issues/4033
